+++ __tests__/stockDetailPage.test.js
@@ -6,6 +6,9 @@ import React from 'react';
 import {render, screen, cleanup} from '@testing-library/react';
 
 import StockDetailPage from '../client/src/components/StockDetailPage.jsx';
+import Graph from '../client/src/components/StockDetailPage/Graph.jsx';
+import {formatAPIData} from '../client/src/components/StockDetailPage/helpers/formatAPIData.js';
+import {getGraphOptions} from '../client/src/components/StockDetailPage/helpers/getGraphOptions.js';
 const {sampleData} = require('../sampleData/sampleData.js');
 
 afterEach(() => {
@@ -42,4 +45,37 @@ test('StockDetailPage renders buy button when user has enough cash balance', () 
   };
   render(<StockDetailPage stockSelected={stockSelected} user={user}/>);
   expect(screen.getByText('Buy')).toBeInTheDocument();
+});
+test('StockDetailPage renders sell button when user owns at least 1 share of the stock', () => {
+  const stockSelected = {
+    name: 'Tesla',
+    symbol: 'TSLA',
+    price: 100,
+    data: sampleData.data
+  };
+  const user = {
+    cashBalance: 100,
+    userPortfolio: [
+      {
+        stockName: 'Tesla',
+        sharesOwned: 1
+      }
+    ]
+  };
+  render(<StockDetailPage stockSelected={stockSelected} user={user}/>);
+  expect(screen.getByText('Sell')).toBeInTheDocument();
+});
+
+test('Graph renders when data is available', () => {
+  const formattedData = formatAPIData(sampleData.data);
+  render(<Graph data={formattedData}/>);
+  expect(screen.getByRole('graph')).toBeInTheDocument();
+});
+
+test('getGraphOptions function returns an option with the correct shape', () => {
+  const formattedData = formatAPIData(sampleData.data);
+  const options = getGraphOptions(formattedData);
+  console.log(options.series[0].data.length);
+  expect(options.series[0].data.length).toBe(formattedData.ohlc.length);
+  expect(options.series[1].data.length).toBe(formattedData.volume.length);
 });
